+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - `Slkecho::SlackApiResponseErorr` to `Slkecho::SlackApiResultError`
 - feat: Improve error message of Slack API request
 - feat: Remove channel validation
+- chore: Update options description
 
 ## [v1.4.0] - 2024-01-09