+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+## [v2.0.0] - 2024-01-12
+
 - feat: Remove --subject option
 - feat: Rename --mention option to --mention-by-email option
 - feat: Rename errors